📄 Commit Guide: Added instructions
Just added some dry instructions on how to commit commits with the commit style. It goes over creating a branch commiting and creating a PR, though it is by no means a step by step guide (missing pushing and adding which is assumed git knoledge).

Mostly for myself but it'll be helpfull to new members, considering I fucked it up once while trying to add this.